feat(ark-client): client to use user-defined timeout when sending out requests
BREAKING CHANGES:

*ark-core:**
- Function signature changed: removing callback from
`history::generate_outgoing_vtxo_transaction_history` and return an Iterator of relevant
transactions. Give the caller more flexibility when generating the list of outgoing
vtxo tx history, so it can handle errors and async calls better when it needs to retrieve
the vtxo from the operator based on the obtained outpoints.

**ark-client:**
- Method signature changed: `connect()` → `connect(timeout)`. The `OfflineClient::connect`
method now expects a `Duration` to be used as the timeout whenever it sends requests to
the Ark operator or Bitcoin Core RPC server.