refactor(ark-core): removing callback from history::generate_outgoing_vtxo_transaction_history signature
BREAKING CHANGES:

**ark-core:**
- Function signature changed: removing callback from
`history::generate_outgoing_vtxo_transaction_history` and return an Iterator of outgoing
transactions. Give the caller more flexibility when generating the list of outgoing
vtxo tx history, so it can handle errors and async calls better when it needs to retrieve
the vtxo from the Operator based on the obtained outpoints.

Co-authored-by: Lucas Soriano <lucas_soriano@fastmail.com>