feat: rework rest client
We use `@openapitools/openapi-generator-cli` to generate a client within `ark-rest`. We want to be able to convert between the autogenerated files and ark-core files, because of this, all files will end up in `ark-rest` directly.

NOTE:
The downside of this approach is that on re-generation certain files such as Cargo.toml, lib.rs, README.md, etc. will be overwritten. One has to be careful when doing so.

We also have a wasm test which runs the resulting file in a headless-browser. Tests are ignored by default as we need arkd running.
To run the wasm tests call `just test-wasm`