ci: move clippy step as a final touch step
Sometimes it is nice to have a fixup commit and be able to
run the tests without having to pass clippy or other repo rules
passing before running the e2e tests.

In other works clippy is a final touch check from the code
prospective, so probably make sense move to the end.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>