feat: Add APIs and CLI commands to settle selectively
Before, settling using `ark-client` and `ark-client-sample` involved
every single VTXO that could be settled.

Settling will eventually have a cost associated with it. Also, only
certain types of VTXOs really need settlement: VTXOs which are close
to expiry and pre-confirmed VTXOs (in some cases).

With this patch we introduce a `settle_vtxos` API in `ark-client`
which allows consumers to specify which VTXOs to settle.

To demonstrate the usefulness of this change, we introduce CLI
commands in `ark-client-sample` to list VTXOs (to identify good
candidates for settlement) and to settle specific VTXOs.

Finally, we combine all this work into a single, self-explanatory
justfile command:

```
just ark-sample settle-recoverable alice
```